feat(repo): fix circular dep
metriport/metriport-internal#1040

Ref: ENG-576
