+++ packages/mllp-server/src/parsing.ts
@@ -1,0 +1,56 @@
+import * as dotenv from "dotenv";
+dotenv.config();
+
+import { Hl7Message } from "@medplum/core";
+import { getSendingApplication } from "@metriport/core/command/hl7v2-subscriptions/hl7v2-to-fhir-conversion/msh";
+import {
+  createUnpackPidFailureFileKey,
+  getCxIdAndPatientIdOrFail,
+  getOptionalValueFromMessage,
+} from "@metriport/core/command/hl7v2-subscriptions/hl7v2-to-fhir-conversion/shared";
+import { utcifyHl7Message } from "@metriport/core/external/hl7-notification/datetime";
+import { Logger } from "@metriport/core/util/log";
+import { buildDayjs, ISO_DATE_TIME } from "@metriport/shared/common/date";
+import { Config } from "./config";
+import { asString, bucketName, s3Utils } from "./utils";
+
+const hieTimezoneDictionary = Config.getHieTimezoneDictionary();
+
+export interface ParsedHl7Data {
+  message: Hl7Message;
+  cxId: string;
+  patientId: string;
+}
+
+export async function parseHl7Message(rawMessage: Hl7Message): Promise<ParsedHl7Data> {
+  const sendingApplication = getSendingApplication(rawMessage) ?? "Unknown HIE";
+  const hieTimezone = hieTimezoneDictionary[sendingApplication] ?? "UTC";
+  const message = utcifyHl7Message(rawMessage, hieTimezone);
+
+  const { cxId, patientId } = getCxIdAndPatientIdOrFail(message);
+
+  return {
+    message,
+    cxId,
+    patientId,
+  };
+}
+export async function handleParsingError(rawMessage: Hl7Message, logger: Logger) {
+  const { log } = logger;
+
+  const fileKey = createUnpackPidFailureFileKey({
+    rawPtIdentifier: getOptionalValueFromMessage(rawMessage, "PID", 3, 1) ?? "unknown-patient",
+    rawTimestamp:
+      getOptionalValueFromMessage(rawMessage, "MSH", 7, 1) ?? buildDayjs().format(ISO_DATE_TIME),
+    messageCode: getOptionalValueFromMessage(rawMessage, "MSH", 9, 1) ?? "UNK",
+    triggerEvent: getOptionalValueFromMessage(rawMessage, "MSH", 9, 2) ?? "UNK",
+  });
+
+  log(`Parsing failed, uploading raw HL7 message to S3 with key: ${fileKey}`);
+  await s3Utils.uploadFile({
+    bucket: bucketName,
+    key: fileKey,
+    file: Buffer.from(asString(rawMessage)),
+    contentType: "text/plain",
+  });
+}
